Note
This code defines a simple CNN with two convolutional layers and one dense layer, which is used to classify the images into different categories. The ImageDataGenerator class is used to read and preprocess the image data, which is in the form of directories containing subdirectories for each class. The flow_from_directory method is used to generate batches of image data for training and testing. The fit method is then used to train the model for a specified number of epochs.